remove tsc requirement and console log
