+++ .travis.yml
@@ -23,7 +23,6 @@ script:
   - docker-compose exec marketjs env TRUFFLE_DEVELOP_HOST="$TRUFFLE_DEVELOP_HOST" npm run test:prod
 after_success:
   - docker-compose exec marketjs env TRAVIS=true TRAVIS_JOB_ID="$TRAVIS_JOB_ID" TRAVIS_BRANCH="$TRAVIS_BRANCH" TRAVIS_PULL_REQUEST="$TRAVIS_PULL_REQUEST" npm run report-coverage
-  - tsc tools/update-docs.ts
   - node tools/update-docs.js
 #  - docker-compose exec marketjs npm run deploy-docs
 #  - docker-compose exec marketjs npm run semantic-release
